+++ src/LionWeb.Core/Core/Notification/CompositeNotification.cs
@@ -50,7 +50,8 @@ public record CompositeNotification : IForestNotification, IPartitionNotificatio
     public INotificationId NotificationId { get; set; }
 
     /// <inheritdoc />
-    public HashSet<IReadableNode> AffectedNodes => Parts.SelectMany(p => p.AffectedNodes).ToHashSet();
+    public HashSet<IReadableNode> AffectedNodes =>
+        Parts.SelectMany(p => p.AffectedNodes).ToHashSet();
 
     /// <inheritdoc />
     public IPartitionInstance Partition => null;